Window Position Profile save and restore scripts
I really lik the program DisplayFusion for the window position profiles.  That's really all I would use it for, I would be able to easily save and restore positions and sizes of open programs, so that when I remove my laptop from the dock, and then put it back, I can quickly get my programs back in place... I'm a little OCD you could say...  I liked that I could add a key combination to quickly restore my positions.

Well, unfortunately, the IT team does NOT like this program, or anything else I have tried up to this point as an alternative such as AquaSnap and quite a few others.  I don't spend a lot of time arguing with them I just remove the program that they ask and move on about my business.

I plan to build a stand alone app in maybe .NET or something that will be visual like DisplayFusion offers for managing and restoring window positions, but I'm in a hurry at the moment and just need something that works for now.  I managed to find a powershell script online, unfortunately I do not recall exactly where the original script came from, that would allow me to record the process.  In theory, this worked for most things, but when you are running multiple windows of the same program, it would get confused and sometimes either not restore them, or put them in the incorrect position.

So I made some adjustments to the original code so that it will look instead at the window title and save the information and then restore from there.  It's primitive, but it seems to work for what I need right now.  But I wanted to also share it out there incase someone is in a similar position where IT keeps gunning for them.

Right now you have to manually type in the names of the windows into the save script, but I hope to make adjustments so that there will be a choice box to choose open apps instead.  But for now, I want to share what I have with the world.